address comments

+++ runtime/near-vm-runner/src/imports.rs
@@ -108,7 +108,7 @@ imports! {
     msg_len: u64,
     msg_ptr: u64,
     pub_key_len: u64,
-    pub_key_ptr: u64] -> [u32]>,
+    pub_key_ptr: u64] -> [u64]>,
     #[MathExtension] ripemd160<[value_len: u64, value_ptr: u64, register_id: u64] -> []>,
     #[MathExtension] ecrecover<[hash_len: u64, hash_ptr: u64, sign_len: u64, sig_ptr: u64, v: u64, malleability_flag: u64, register_id: u64] -> [u64]>,
     // #####################
